backend/requirements.txt
fastapi
uvicorn[standard]
playwright
google-genai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
anthropic
python-dotenv
pydantic
httpx
sqlalchemy
